added timeline to world

+++ engine/probability_machine/random.go
@@ -103,11 +103,11 @@ func randFloat64NormInRange(min, max, stdDev, mean float64, count int) (float64,
 	return r / multiplier, nil
 }
 
-func RandomValueOfSliceNorm[T interface{}](meanIndex int, sl []T) (T, error) {
+func RandomValueOfSliceNorm[T interface{}](meanIndex float64, sl []T) (T, error) {
 	var zero T
 
-	if meanIndex >= len(sl) {
-		return zero, wrapped_error.NewInternalServerError(nil, fmt.Sprintf("can not get random value with norm (mean_index=%d, slice length=%d)", meanIndex, len(sl)))
+	if meanIndex >= float64(len(sl)) {
+		return zero, wrapped_error.NewInternalServerError(nil, fmt.Sprintf("can not get random value with norm (mean_index=%f, slice length=%d)", meanIndex, len(sl)))
 	}
 	indexF, err := RandFloat64NormInRange(0, float64(len(sl)-1), 1, float64(meanIndex))
 	if err != nil {
